🧹 Refactor DashboardPage into smaller components
